fix(llm): move the cloud agnostic methods of the controller to the base
